Add files via upload
Sample designs of a multi-signature smart contract that can manage treasury assets (ETH/ERC20, but ERC721 and ERC1155 aren't hard to implement either), as well as a variety of scrapped designs that were written along the way to the final V3 design. These contracts demonstrate understanding of Solidity syntax, design considerations, and the NatSpec format for commenting Solidity contracts.